Retrieve BIP scores for a list of DOIs

This function takes a list of DOIs, normalizes them, queries the BIP API to retrieve scores, and returns a data frame with bibliometric indicators.

Usage

get_BIP(dois, verbose = FALSE)

Arguments

dois

A character vector of DOIs

verbose

Show diagnostic information?

Value

A data frame with bibliometric indicators for the provided DOIs

Examples

dois <- c("https://doi.org/10.1080/00223891.2020.1726936", "10.1016/j.jrp.2013.05.009")
get_BIP(dois)
#>                                             doi      attrank tar_ram
#> 1 https://doi.org/10.1080/00223891.2020.1726936 1.710533e-08      NA
#> 2     https://doi.org/10.1016/j.jrp.2013.05.009 8.816282e-07      NA
#>       pagerank three_year_cc   cc pop_class inf_class imp_class msg
#> 1 3.889392e-09            19   22        C4        C5        C4  NA
#> 2 6.673346e-08            93 1452        C1        C2        C3  NA