Get a reference set for field normalized citation rates

This function downloads a large set of publications from a certain field and certain years. Retrieving 10,000 documents from the API takes about 1-2 min. Ideally, this should be cached locally.

Usage

get_reference_set(
  from_year,
  to_year = NA,
  n_per_year = 10000,
  concept.id = "C15744967",
  verbose = TRUE
)

Arguments

from_year

The start year of the range

to_year

The end year of the range (optional, defaults to from_year)

n_per_year

The number of documents to retrieve per requested year (optional, defaults to 10000). Values larger than 10000 are possible (they are split up to multiple OA requests).

concept.id

A vector of concept.ids to search for (optional, defaults to "C15744967", i.e. "Psychology")

verbose

Show OA API progress?

Value

A data frame containing the document id, year, cited_by_count, and number of authors of the retrieved documents

Examples

# Get reference set for "Psychology" for multiple years (small n here for demo)
psych_ref <- get_reference_set(
   from_year = 2013, to_year = 2020,
   n_per_year = 100, concept.id = "C15744967"
 )
#> [1] "2023-05-24 09:22:44.669528: Retrieving documents for 2013 ..."
#> Requesting url: https://api.openalex.org/works?filter=sample%3A100%2Cseed%3A201301%2Cconcept.id%3AC15744967%2Ctype%3Ajournal-article%2Cfrom_publication_date%3A2013-01-01%2Cto_publication_date%3A2013-12-31%2Cis_paratext%3Afalse%2Cis_retracted%3Afalse%2Cauthors_count%3A%3E0%2Chas_doi%3Atrue
#> Error: OpenAlex API request failed [403]
#> Invalid query parameters error.
#> <sample is not a valid field. Valid fields are underscore or hyphenated versions of: abstract.search, apc_payment.currency, apc_payment.price, apc_payment.price_usd, apc_payment.provenance, author.id, author.orcid, authors_count, authorships.author.id, authorships.author.orcid, authorships.institutions.continent, authorships.institutions.country_code, authorships.institutions.id, authorships.institutions.is_global_south, authorships.institutions.ror, authorships.institutions.type, authorships.is_corresponding, best_oa_location.is_oa, best_oa_location.license, best_oa_location.source.host_organization, best_oa_location.source.host_organization_lineage, best_oa_location.source.id, best_oa_location.source.issn, best_oa_location.source.type, best_oa_location.venue.issn, best_oa_location.venue.type, best_oa_location.version, best_open_version, cited_by, cited_by_count, cites, concept.id, concepts.id, concepts.wikidata, concepts_count, corresponding_author_ids, corresponding_institution_ids, default.search, display_name, display_name.search, doi, doi_starts_with, from_created_date, from_publication_date, fulltext.search, grants.award_id, grants.funder, has_abstract, has_doi, has_fulltext, has_ngrams, has_oa_accepted_or_published_version, has_oa_submitted_version, has_orcid, has_pdf_url, has_pmcid, has_pmid, has_raw_affiliation_string, has_references, host_venue.id, host_venue.publisher, ids.mag, ids.openalex, ids.pmcid, ids.pmid, institution.id, institutions.continent, institutions.country_code, institutions.id, institutions.is_global_south, institutions.ror, institutions.type, is_corresponding, is_oa, is_paratext, is_retracted, journal, language, locations.is_oa, locations.license, locations.source.host_institution_lineage, locations.source.host_organization, locations.source.host_organization_lineage, locations.source.id, locations.source.issn, locations.source.publisher_lineage, locations.source.type, locations.venue.issn, locations.venue.type, locations.version, locations_count, mag, oa_status, open_access.any_repository_has_fulltext, open_access.is_oa, open_access.oa_status, openalex, openalex_id, pmcid, pmid, primary_location.is_oa, primary_location.license, primary_location.source.has_issn, primary_location.source.host_institution_lineage, primary_location.source.host_organization, primary_location.source.host_organization_lineage, primary_location.source.id, primary_location.source.issn, primary_location.source.publisher_lineage, primary_location.source.type, primary_location.venue.has_issn, primary_location.version, publication_date, publication_year, raw_affiliation_string.search, referenced_works, related_to, repository, title.search, to_publication_date, type, version>