Compute h-index and hIa-index

TODO: Could use a time-window (currently, it uses all citations since publication).

Usage

h_index(
  search = NULL,
  display_name = NULL,
  author.id = NULL,
  ORCID = NULL,
  first_pub_year = NA,
  academic_age_bonus = 0
)

Arguments

search

A character string with the author's name for searching in the OpenAlex API. Allows fuzzy search (e.g. with or without middle initial).

display_name

A character vector with the exact author's names for searching in the OpenAlex API. You cvan provide multiple variants in a vector.

author.id

A character vector with OpenAlex author ID(s) (i.e., an ID starting with "A.......") for searching in the OpenAlex API.

ORCID

A character vector with ORCID IDs (either as URI or as the number quartet string).

first_pub_year

An integer with the year of the author's first publication. If not provided, it is estimated from the retrieved publications.

academic_age_bonus

Years that are subtracted from active academic life (e.g. due to child care). For example, SNF subtracts 1.5 years per child (see https://www.snf.ch/en/cciM9NWuvhOVKRxv/news/news-200803-career-funding-the-researchers-overall-performance-counts).

Value

A list containing the net academic age, h-index and hIa-index of the author.

Details

Computes the h-index and hIa-index of an author. You need to provide exactly one of the parameters search, display_name, author.id, or ORCID:

search allows fuzzy search of the author's name (e.g. with or without middle initial). Might return too many matches

display_name requires an exact match but allows to provide a vector with multiple versions of a name (e.g., c("Hans Müller", "Hans Mueller", "Hans J. Muller"))

author.id is the OpenAlex ID for authors.

ORCID is the ORCID (either as link or as the number quartet string)

Often the same author is listed under different IDs in the OpenAlex database, therefore these are merged by the function. But what if the function includes wrong authors? Then you need to manually screen the table with author names, record the IDs from the correct aliases of the target person (starting with "A.......") and enter these author IDs into the author.id parameter.

Examples

h_index(search = "Felix Schönbrodt")
#> A unique author record has been found: 
#> 
#> # A tibble: 1 × 3
#>   display_name        works_count cited_by_count
#>   <chr>                     <int>          <int>
#> 1 Felix D. Schönbrodt         252          12348
#> Warning: 
#> The following work(s) have truncated lists of authors: W4407393904, W3216869777.
#> Query each work separately by its identifier to get full list of authors.
#> For example:
#>   lapply(c("W4407393904", "W3216869777"), \(x) oa_fetch(identifier = x))
#> Details at https://docs.openalex.org/api-entities/authors/limitations.
#> Computation of academic age (relative to first publication):
#> Retrieving year of first publication from data: 2007. 
#> 
#> The first publication is: 'Interview mit Prof. Dr. Karl Grammer', published in 'Zeitschrift für Medienpsychologie'. It has the top L0 concept 'Philosophy'. 
#> 
#> Is that plausible? If not, provide `first_pub_year` as a parameter. You can see all works of that author with the following command:
#> 
#> oa_fetch(entity = 'works', author.id = 'https://openalex.org/A5022479713', is_paratext = F, is_retracted = F, abstract=F) %>% arrange(publication_year)
#> 
#> $net_academic_age
#> [1] 18
#> 
#> $h_index
#> [1] 34
#> 
#> $h_Ia
#> [1] 1.055556
#> 
#> $first_publication
#> # A tibble: 1 × 39
#>   id    title display_name author publication_date so    so_id host_organization
#>   <chr> <chr> <chr>        <list> <chr>            <chr> <chr> <chr>            
#> 1 http… Inte… Interview m… <df>   2007-10-01       Zeit… http… Hogrefe Verlag   
#> # ℹ 31 more variables: issn_l <chr>, url <chr>, pdf_url <chr>, license <chr>,
#> #   version <chr>, first_page <chr>, last_page <chr>, volume <chr>,
#> #   issue <chr>, is_oa <lgl>, is_oa_anywhere <lgl>, oa_status <chr>,
#> #   oa_url <chr>, any_repository_has_fulltext <lgl>, language <chr>,
#> #   grants <list>, cited_by_count <int>, counts_by_year <list>,
#> #   publication_year <int>, cited_by_api_url <chr>, ids <list>, doi <chr>,
#> #   type <chr>, referenced_works <list>, related_works <list>, …
#> 
#> $academic_age_note
#> [1] "Computation of academic age (relative to first publication):\nRetrieving year of first publication from data: 2007. \n\nThe first publication is: 'Interview mit Prof. Dr. Karl Grammer', published in 'Zeitschrift für Medienpsychologie'. It has the top L0 concept 'Philosophy'. \n\nIs that plausible? If not, provide `first_pub_year` as a parameter. You can see all works of that author with the following command:\n\noa_fetch(entity = 'works', author.id = 'https://openalex.org/A5022479713', is_paratext = F, is_retracted = F, abstract=F) %>% arrange(publication_year)\n\n"
#> 
h_index(ORCID = "https://orcid.org/0000-0002-8282-3910")
#> A unique author record has been found: 
#> 
#> # A tibble: 1 × 3
#>   display_name        works_count cited_by_count
#>   <chr>                     <int>          <int>
#> 1 Felix D. Schönbrodt         252          12348
#> Warning: 
#> The following work(s) have truncated lists of authors: W4407393904, W3216869777.
#> Query each work separately by its identifier to get full list of authors.
#> For example:
#>   lapply(c("W4407393904", "W3216869777"), \(x) oa_fetch(identifier = x))
#> Details at https://docs.openalex.org/api-entities/authors/limitations.
#> Computation of academic age (relative to first publication):
#> Retrieving year of first publication from data: 2007. 
#> 
#> The first publication is: 'Interview mit Prof. Dr. Karl Grammer', published in 'Zeitschrift für Medienpsychologie'. It has the top L0 concept 'Philosophy'. 
#> 
#> Is that plausible? If not, provide `first_pub_year` as a parameter. You can see all works of that author with the following command:
#> 
#> oa_fetch(entity = 'works', author.id = 'https://openalex.org/A5022479713', is_paratext = F, is_retracted = F, abstract=F) %>% arrange(publication_year)
#> 
#> $net_academic_age
#> [1] 18
#> 
#> $h_index
#> [1] 34
#> 
#> $h_Ia
#> [1] 1.055556
#> 
#> $first_publication
#> # A tibble: 1 × 39
#>   id    title display_name author publication_date so    so_id host_organization
#>   <chr> <chr> <chr>        <list> <chr>            <chr> <chr> <chr>            
#> 1 http… Inte… Interview m… <df>   2007-10-01       Zeit… http… Hogrefe Verlag   
#> # ℹ 31 more variables: issn_l <chr>, url <chr>, pdf_url <chr>, license <chr>,
#> #   version <chr>, first_page <chr>, last_page <chr>, volume <chr>,
#> #   issue <chr>, is_oa <lgl>, is_oa_anywhere <lgl>, oa_status <chr>,
#> #   oa_url <chr>, any_repository_has_fulltext <lgl>, language <chr>,
#> #   grants <list>, cited_by_count <int>, counts_by_year <list>,
#> #   publication_year <int>, cited_by_api_url <chr>, ids <list>, doi <chr>,
#> #   type <chr>, referenced_works <list>, related_works <list>, …
#> 
#> $academic_age_note
#> [1] "Computation of academic age (relative to first publication):\nRetrieving year of first publication from data: 2007. \n\nThe first publication is: 'Interview mit Prof. Dr. Karl Grammer', published in 'Zeitschrift für Medienpsychologie'. It has the top L0 concept 'Philosophy'. \n\nIs that plausible? If not, provide `first_pub_year` as a parameter. You can see all works of that author with the following command:\n\noa_fetch(entity = 'works', author.id = 'https://openalex.org/A5022479713', is_paratext = F, is_retracted = F, abstract=F) %>% arrange(publication_year)\n\n"
#> 
h_index("Markus Bühner", first_pub_year = 1997) # adjust year of first publication
#> A unique author record has been found: 
#> 
#> # A tibble: 1 × 3
#>   display_name  works_count cited_by_count
#>   <chr>               <int>          <int>
#> 1 Markus Bühner         211           7810
#> $net_academic_age
#> [1] 28
#> 
#> $h_index
#> [1] 41
#> 
#> $h_Ia
#> [1] 0.6071429
#> 
#> $first_publication
#> # A tibble: 1 × 39
#>   id    title display_name author publication_date so    so_id host_organization
#>   <chr> <chr> <chr>        <list> <chr>            <chr> <chr> <chr>            
#> 1 http… Attr… Attribution… <df>   1998-01-01       Jour… http… American Psychol…
#> # ℹ 31 more variables: issn_l <chr>, url <chr>, pdf_url <chr>, license <chr>,
#> #   version <chr>, first_page <chr>, last_page <chr>, volume <chr>,
#> #   issue <chr>, is_oa <lgl>, is_oa_anywhere <lgl>, oa_status <chr>,
#> #   oa_url <chr>, any_repository_has_fulltext <lgl>, language <chr>,
#> #   grants <list>, cited_by_count <int>, counts_by_year <list>,
#> #   publication_year <int>, cited_by_api_url <chr>, ids <list>, doi <chr>,
#> #   type <chr>, referenced_works <list>, related_works <list>, …
#> 
#> $academic_age_note
#> [1] ""
#>