Get network analysis of co-authorships for a given author or set of works

This function performs a network analysis for a given author or set of works, providing indicators for the degree of international collaborations.

Usage

get_network(author.id, doi = NA, works = NA, min_edges = 2, verbose = TRUE)

Arguments

author.id

The ID of the author for whom the network will be analyzed. Must be provided.

doi

Optional vector of DOIs. If provided, only works with these DOIs will be analyzed.

works

Optional data frame of works (that have already been fetched by oa_fetch). Either provide dois or works.

min_edges

Minimal number of co-authorships. By default set to 2: This way, single co-authorship connections are ignored (e.g., in ManyLabs-style papers, this can lead to a strong inflation of co-authorship connections.)

verbose

Whether to display the result string (default is TRUE).

Value

A list containing various results of the network analysis, including unique coauthor edges, counts of international and same-country co-authors, country codes of coauthors, and a result string summarizing the analysis. It also computes the normalized Shannon entropy. It is normalized by dividing it by log_2(n_countries), which bounds its range to [0; 1].

Examples

# Analyze the network of a specific author
get_network(author.id = 'https://openalex.org/A5022479713')
#> Requesting url: https://api.openalex.org/works?filter=author.id%3Ahttps%3A%2F%2Fopenalex.org%2FA5022479713%2Cis_paratext%3Afalse%2Cis_retracted%3Afalse
#> Getting 2 pages of results with a total of 242 records...
#> Warning: 
#> The following work(s) have truncated lists of authors: W3216869777.
#> Query each work separately by its identifier to get full list of authors.
#> For example:
#>   lapply(c("W3216869777"), \(x) oa_fetch(identifier = x))
#> Details at https://docs.openalex.org/api-entities/authors/limitations.
#> [1] "Using the provided works."
#> [1] "346 unique co-authors (n>=2 joint papers). 71% from 26 international countries, 29% from the same country. The normalized entropy (ranging from 0=only one country to 1=even distribution among all countries) is 0.65. The 4 countries with the most coauthors are DE, US, NL, GB"
#> $unique_coauthor_edges
#> # A tibble: 604 × 13
#>    au_id               au_display_name au_orcid author_position is_corresponding
#>    <chr>               <chr>           <chr>    <chr>           <lgl>           
#>  1 https://openalex.o… Birk Hagemeyer  NA       middle          FALSE           
#>  2 https://openalex.o… Steffen Nestler https:/… last            FALSE           
#>  3 https://openalex.o… Caroline Zygar… https:/… first           FALSE           
#>  4 https://openalex.o… Mitja D. Back   https:/… first           TRUE            
#>  5 https://openalex.o… Sarah Humberg   https:/… middle          FALSE           
#>  6 https://openalex.o… Jaap J. A. Den… https:/… middle          FALSE           
#>  7 https://openalex.o… Lena Schiestel  NA       middle          FALSE           
#>  8 https://openalex.o… Eric‐Jan Wagen… https:/… middle          FALSE           
#>  9 https://openalex.o… Michael Dufner  https:/… first           FALSE           
#> 10 https://openalex.o… Marco Perugini  https:/… middle          FALSE           
#> # ℹ 594 more rows
#> # ℹ 8 more variables: au_affiliation_raw <chr>, institution_id <chr>,
#> #   institution_display_name <chr>, institution_ror <chr>,
#> #   institution_country_code <chr>, institution_type <chr>,
#> #   institution_lineage <chr>, n_coauthorships <int>
#> 
#> $n_coauthors_international
#> [1] 217
#> 
#> $n_coauthors_same_country
#> [1] 87
#> 
#> $entropy_normalized
#> [1] 0.6454006
#> 
#> $country_codes_repeated
#> # A tibble: 27 × 2
#>    institution_country_code     n
#>    <chr>                    <int>
#>  1 DE                          87
#>  2 US                          86
#>  3 NL                          39
#>  4 GB                          27
#>  5 IT                          10
#>  6 CH                           8
#>  7 SE                           7
#>  8 AT                           5
#>  9 ES                           4
#> 10 GR                           4
#> # ℹ 17 more rows
#> 
#> $result_string
#> [1] "346 unique co-authors (n>=2 joint papers). 71% from 26 international countries, 29% from the same country. The normalized entropy (ranging from 0=only one country to 1=even distribution among all countries) is 0.65. The 4 countries with the most coauthors are DE, US, NL, GB"
#> 
# Analyze the network of a specific author for selected works
get_network(author.id = 'https://openalex.org/A5022479713', min_edges=1,
    doi = c(
        'https://doi.org/10.1037/pspp0000428',
        'https://doi.org/10.1017/S0033291722003294',
        'https://doi.org/10.5964/ps.6029',
        'https://doi.org/10.1146/annurev-psych-020821-114157'))
#> Requesting url: https://api.openalex.org/works?filter=doi%3Ahttps%3A%2F%2Fdoi.org%2F10.1037%2Fpspp0000428%7Chttps%3A%2F%2Fdoi.org%2F10.1017%2FS0033291722003294%7Chttps%3A%2F%2Fdoi.org%2F10.5964%2Fps.6029%7Chttps%3A%2F%2Fdoi.org%2F10.1146%2Fannurev-psych-020821-114157%2Cis_paratext%3Afalse%2Cis_retracted%3Afalse
#> Getting 1 page of results with a total of 4 records...
#> [1] "35 unique co-authors (n>=1 joint papers). 52% from 4 international countries, 48% from the same country. The normalized entropy (ranging from 0=only one country to 1=even distribution among all countries) is 0.81. The 4 countries with the most coauthors are DE, US, NL, AU"
#> $unique_coauthor_edges
#> # A tibble: 35 × 13
#>    au_id               au_display_name au_orcid author_position is_corresponding
#>    <chr>               <chr>           <chr>    <chr>           <lgl>           
#>  1 https://openalex.o… Büsra Bektas    NA       middle          FALSE           
#>  2 https://openalex.o… André Wannemül… https:/… middle          FALSE           
#>  3 https://openalex.o… Anne M. Scheel  https:/… middle          FALSE           
#>  4 https://openalex.o… Fiona Fidler    https:/… middle          FALSE           
#>  5 https://openalex.o… Simine Vazire   https:/… last            FALSE           
#>  6 https://openalex.o… Roos Hutteman   NA       middle          FALSE           
#>  7 https://openalex.o… Isabel Thielma… https:/… middle          FALSE           
#>  8 https://openalex.o… Katherine S. C… https:/… middle          FALSE           
#>  9 https://openalex.o… Anna Dreber     https:/… middle          FALSE           
#> 10 https://openalex.o… Melissa Kline … NA       middle          FALSE           
#> # ℹ 25 more rows
#> # ℹ 8 more variables: au_affiliation_raw <chr>, institution_id <chr>,
#> #   institution_display_name <chr>, institution_ror <chr>,
#> #   institution_country_code <chr>, institution_type <chr>,
#> #   institution_lineage <chr>, n_coauthorships <int>
#> 
#> $n_coauthors_international
#> [1] 14
#> 
#> $n_coauthors_same_country
#> [1] 13
#> 
#> $entropy_normalized
#> [1] 0.8075147
#> 
#> $country_codes_repeated
#> # A tibble: 5 × 2
#>   institution_country_code     n
#>   <chr>                    <int>
#> 1 DE                          13
#> 2 US                           7
#> 3 NL                           4
#> 4 AU                           2
#> 5 SE                           1
#> 
#> $result_string
#> [1] "35 unique co-authors (n>=1 joint papers). 52% from 4 international countries, 48% from the same country. The normalized entropy (ranging from 0=only one country to 1=even distribution among all countries) is 0.81. The 4 countries with the most coauthors are DE, US, NL, AU"
#>