Get network analysis of co-authorships for a given author or set of works

This function performs a network analysis for a given author or set of works, providing indicators for the degree of international collaborations.

Usage

get_network(author.id, doi = NA, works = NA, min_edges = 2, verbose = TRUE)

Arguments

author.id

The ID of the author for whom the network will be analyzed. Must be provided.

doi

Optional vector of DOIs. If provided, only works with these DOIs will be analyzed.

works

Optional data frame of works (that have already been fetched by oa_fetch). Either provide dois or works.

min_edges

Minimal number of co-authorships. By default set to 2: This way, single co-authorship connections are ignored (e.g., in ManyLabs-style papers, this can lead to a strong inflation of co-authorship connections.)

verbose

Whether to display the result string (default is TRUE).

Value

A list containing various results of the network analysis, including unique coauthor edges, counts of international and same-country co-authors, country codes of coauthors, and a result string summarizing the analysis. It also computes the normalized Shannon entropy. It is normalized by dividing it by log_2(n_countries), which bounds its range to [0; 1].

Examples

# Analyze the network of a specific author
get_network(author.id = 'https://openalex.org/A5022479713')
#> Requesting url: https://api.openalex.org/works?filter=author.id%3Ahttps%3A%2F%2Fopenalex.org%2FA5022479713%2Cis_paratext%3Afalse%2Cis_retracted%3Afalse
#> Getting 2 pages of results with a total of 246 records...
#> Warning: 
#> The following work(s) have truncated lists of authors: W3216869777.
#> Query each work separately by its identifier to get full list of authors.
#> For example:
#>   lapply(c("W3216869777"), \(x) oa_fetch(identifier = x))
#> Details at https://docs.openalex.org/api-entities/authors/limitations.
#> [1] "Using the provided works."
#> [1] "341 unique co-authors (n>=2 joint papers). 73% from 27 international countries, 27% from the same country. The normalized entropy (ranging from 0=only one country to 1=even distribution among all countries) is 0.65. The 4 countries with the most coauthors are US, DE, NL, GB"
#> $unique_coauthor_edges
#> # A tibble: 583 × 3
#>    au_id                            institution_country_code n_coauthorships
#>    <chr>                            <chr>                              <int>
#>  1 https://openalex.org/A5039784616 DE                                    24
#>  2 https://openalex.org/A5063079579 NL                                    24
#>  3 https://openalex.org/A5056348040 DE                                    22
#>  4 https://openalex.org/A5055189096 DE                                    19
#>  5 https://openalex.org/A5030640837 DE                                    18
#>  6 https://openalex.org/A5083504666 DE                                    18
#>  7 https://openalex.org/A5052525257 DE                                    17
#>  8 https://openalex.org/A5062095202 DE                                    17
#>  9 https://openalex.org/A5000733234 DE                                    13
#> 10 https://openalex.org/A5067075596 NA                                    13
#> # ℹ 573 more rows
#> 
#> $n_coauthors_international
#> [1] 230
#> 
#> $n_coauthors_same_country
#> [1] 83
#> 
#> $entropy_normalized
#> [1] 0.6497213
#> 
#> $country_codes_repeated
#> # A tibble: 28 × 2
#>    institution_country_code     n
#>    <chr>                    <int>
#>  1 US                          89
#>  2 DE                          83
#>  3 NL                          45
#>  4 GB                          27
#>  5 IT                          10
#>  6 CH                           8
#>  7 SE                           7
#>  8 AT                           6
#>  9 ES                           4
#> 10 GR                           4
#> # ℹ 18 more rows
#> 
#> $result_string
#> [1] "341 unique co-authors (n>=2 joint papers). 73% from 27 international countries, 27% from the same country. The normalized entropy (ranging from 0=only one country to 1=even distribution among all countries) is 0.65. The 4 countries with the most coauthors are US, DE, NL, GB"
#> 
# Analyze the network of a specific author for selected works
get_network(author.id = 'https://openalex.org/A5022479713', min_edges=1,
    doi = c(
        'https://doi.org/10.1037/pspp0000428',
        'https://doi.org/10.1017/S0033291722003294',
        'https://doi.org/10.5964/ps.6029',
        'https://doi.org/10.1146/annurev-psych-020821-114157'))
#> Requesting url: https://api.openalex.org/works?filter=doi%3Ahttps%3A%2F%2Fdoi.org%2F10.1037%2Fpspp0000428%7Chttps%3A%2F%2Fdoi.org%2F10.1017%2FS0033291722003294%7Chttps%3A%2F%2Fdoi.org%2F10.5964%2Fps.6029%7Chttps%3A%2F%2Fdoi.org%2F10.1146%2Fannurev-psych-020821-114157%2Cis_paratext%3Afalse%2Cis_retracted%3Afalse
#> Getting 1 page of results with a total of 4 records...
#> [1] "35 unique co-authors (n>=1 joint papers). 52% from 4 international countries, 48% from the same country. The normalized entropy (ranging from 0=only one country to 1=even distribution among all countries) is 0.81. The 4 countries with the most coauthors are DE, US, NL, AU"
#> $unique_coauthor_edges
#> # A tibble: 35 × 3
#>    au_id                            institution_country_code n_coauthorships
#>    <chr>                            <chr>                              <int>
#>  1 https://openalex.org/A5001668502 DE                                     1
#>  2 https://openalex.org/A5003489815 DE                                     1
#>  3 https://openalex.org/A5004866926 NL                                     1
#>  4 https://openalex.org/A5006406097 AU                                     1
#>  5 https://openalex.org/A5011404541 AU                                     1
#>  6 https://openalex.org/A5015649291 NA                                     1
#>  7 https://openalex.org/A5017494791 DE                                     1
#>  8 https://openalex.org/A5021314811 SE                                     1
#>  9 https://openalex.org/A5027185097 US                                     1
#> 10 https://openalex.org/A5027692708 DE                                     1
#> # ℹ 25 more rows
#> 
#> $n_coauthors_international
#> [1] 14
#> 
#> $n_coauthors_same_country
#> [1] 13
#> 
#> $entropy_normalized
#> [1] 0.8075147
#> 
#> $country_codes_repeated
#> # A tibble: 5 × 2
#>   institution_country_code     n
#>   <chr>                    <int>
#> 1 DE                          13
#> 2 US                           7
#> 3 NL                           4
#> 4 AU                           2
#> 5 SE                           1
#> 
#> $result_string
#> [1] "35 unique co-authors (n>=1 joint papers). 52% from 4 international countries, 48% from the same country. The normalized entropy (ranging from 0=only one country to 1=even distribution among all countries) is 0.81. The 4 countries with the most coauthors are DE, US, NL, AU"
#>