Get network analysis of co-authorships for a given author or set of works

This function performs a network analysis for a given author or set of works, providing indicators for the degree of international collaborations.

Usage

get_network(
  author.id,
  doi = NA,
  works = NA,
  min_coauthorships = 2,
  verbose = TRUE
)

Arguments

author.id

The ID of the author for whom the network will be analyzed. Must be provided.

doi

Optional vector of DOIs. If provided, only works with these DOIs will be analyzed.

works

Optional data frame of works (that have already been fetched by oa_fetch). Either provide dois or works.

min_coauthorships

Minimal number of co-authorships. By default set to 2: This way, single co-authorship connections are ignored (e.g., in ManyLabs-style papers, this can lead to a strong inflation of co-authorship connections.)

verbose

Whether to display the result string (default is TRUE).

Value

A list containing various results of the network analysis, including unique coauthor edges, counts of international and same-country co-authors, country codes of coauthors, and a result string summarizing the analysis. It also computes the normalized Shannon entropy. It is normalized by dividing it by log_2(n_countries), which bounds its range to [0; 1].

Examples

# Analyze the network of a specific author
get_network(author.id = 'https://openalex.org/A5022479713')
#> Requesting url: https://api.openalex.org/works?filter=author.id%3Ahttps%3A%2F%2Fopenalex.org%2FA5022479713%2Cis_paratext%3Afalse%2Cis_retracted%3Afalse
#> Getting 2 pages of results with a total of 251 records...
#> Warning: 
#> The following work(s) have truncated lists of authors: W4407393904, W3216869777.
#> Query each work separately by its identifier to get full list of authors.
#> For example:
#>   lapply(c("W4407393904", "W3216869777"), \(x) oa_fetch(identifier = x))
#> Details at https://docs.openalex.org/api-entities/authors/limitations.
#> [1] "Using the provided works."
#> [1] "269 unique identifiable co-authors with at least 2 joint papers; 74% from 27 international countries, 26% from the same country. The evenness (ranging from 0=only one country to 1=even distribution among all countries) is 0.77. The 4 countries with the most coauthors are: US (75), DE (71), NL (35), GB (25)."
#> [1] "4 primary fields: Psychology (128), Decision Sciences (43), Computer Science (28), Health Professions (7). The evenness (ranging from 0=only one field to 1=even distribution among all fields) is 0.56."
#> $unique_coauthor_edges
#> # A tibble: 418 × 3
#>    au_id                            institution_country_code n_coauthorships
#>    <chr>                            <chr>                              <int>
#>  1 https://openalex.org/A5056348040 DE                                    18
#>  2 https://openalex.org/A5063079579 NL                                    17
#>  3 https://openalex.org/A5039784616 DE                                    16
#>  4 https://openalex.org/A5052525257 DE                                    13
#>  5 https://openalex.org/A5055189096 DE                                    12
#>  6 https://openalex.org/A5083504666 DE                                    11
#>  7 https://openalex.org/A5031368517 DE                                    10
#>  8 https://openalex.org/A5089676027 DE                                    10
#>  9 https://openalex.org/A5000733234 DE                                     9
#> 10 https://openalex.org/A5004432486 DE                                     9
#> # ℹ 408 more rows
#> 
#> $n_coauthors_international
#> [1] 198
#> 
#> $n_coauthors_same_country
#> [1] 71
#> 
#> $international_evenness
#> [1] 0.769137
#> 
#> $country_codes_repeated
#> # A tibble: 28 × 2
#>    institution_country_code     n
#>    <chr>                    <int>
#>  1 US                          75
#>  2 DE                          71
#>  3 NL                          35
#>  4 GB                          25
#>  5 IT                           9
#>  6 CH                           7
#>  7 SE                           7
#>  8 AT                           4
#>  9 ES                           4
#> 10 GR                           4
#> # ℹ 18 more rows
#> 
#> $internationalization_string
#> [1] "269 unique identifiable co-authors with at least 2 joint papers; 74% from 27 international countries, 26% from the same country. The evenness (ranging from 0=only one country to 1=even distribution among all countries) is 0.77. The 4 countries with the most coauthors are: US (75), DE (71), NL (35), GB (25)."
#> 
#> $interdisc_evenness
#> [1] 0.5630629
#> 
#> $primary_fields_tab
#>                                   primary_field   n
#> 1                                    Psychology 128
#> 2                             Decision Sciences  43
#> 3                              Computer Science  28
#> 4                            Health Professions   7
#> 5                           Arts and Humanities   4
#> 6                                   Mathematics   4
#> 7                                      Medicine   4
#> 8  Biochemistry, Genetics and Molecular Biology   3
#> 9           Business, Management and Accounting   2
#> 10                                  Engineering   2
#> 11                                 Neuroscience   2
#> 12         Agricultural and Biological Sciences   1
#> 13          Economics, Econometrics and Finance   1
#> 14                        Environmental Science   1
#> 15                        Physics and Astronomy   1
#> 
#> $primary_fields_tab_reduced
#>        primary_field   n
#> 1         Psychology 128
#> 2  Decision Sciences  43
#> 3   Computer Science  28
#> 4 Health Professions   7
#> 
#> $subfields_tab
#>                                                subfield  n
#> 1                 Experimental and Cognitive Psychology 84
#> 2                                     Social Psychology 82
#> 3               Statistics, Probability and Uncertainty 43
#> 4                       Sociology and Political Science 41
#> 5                                   Clinical Psychology 40
#> 6                                    Applied Psychology 38
#> 7                                   Information Systems 25
#> 8                    Information Systems and Management 20
#> 9            Management Science and Operations Research 17
#> 10                           Statistics and Probability 13
#> 11                               Cognitive Neuroscience 12
#> 12                 Computer Networks and Communications 12
#> 13                              Artificial Intelligence  9
#> 14                                   General Psychology  9
#> 15                                               Health  7
#> 16                           General Health Professions  6
#> 17                           Economics and Econometrics  5
#> 18 Public Health, Environmental and Occupational Health  5
#> 19                    Statistical and Nonlinear Physics  5
#> 20                        Computer Science Applications  4
#> 21             Developmental and Educational Psychology  4
#> 22                                           Demography  3
#> 23                                         Food Science  3
#> 24                    History and Philosophy of Science  3
#> 25                                    Molecular Biology  3
#> 26                                      Safety Research  3
#> 27                               Biomedical Engineering  2
#> 28              Computer Vision and Pattern Recognition  2
#> 29                      Control and Systems Engineering  2
#> 30                                       Gender Studies  2
#> 31                              General Social Sciences  2
#> 32                                             Genetics  2
#> 33                       Management Information Systems  2
#> 34              Management of Technology and Innovation  2
#> 35                                  Medical Terminology  2
#> 36                                            Museology  2
#> 37                                             Software  2
#> 38                              Strategy and Management  2
#> 39                               Automotive Engineering  1
#> 40                              Behavioral Neuroscience  1
#> 41                                         Biochemistry  1
#> 42               Cardiology and Cardiovascular Medicine  1
#> 43                                        Communication  1
#> 44          Computer Graphics and Computer-Aided Design  1
#> 45         Ecology, Evolution, Behavior and Systematics  1
#> 46                                            Education  1
#> 47                            General Decision Sciences  1
#> 48                                  Infectious Diseases  1
#> 49                             Language and Linguistics  1
#> 50                       Literature and Literary Theory  1
#> 51               Management, Monitoring, Policy and Law  1
#> 52                                         Pharmacology  1
#> 53        Political Science and International Relations  1
#> 54                   Pulmonary and Respiratory Medicine  1
#> 55              Radiology, Nuclear Medicine and Imaging  1
#> 56                                Reproductive Medicine  1
#> 57                                    Signal Processing  1
#> 
#> $topics_tab
#>                                                                  topic  n
#> 1                                        Mental Health Research Topics 49
#> 2                                 Attachment and Relationship Dynamics 25
#> 3                                 Meta-analysis and systematic reviews 23
#> 4                                  Behavioral Health and Interventions 22
#> 5                                      Cultural Differences and Values 20
#> 6                                   Research Data Management Practices 20
#> 7                           Evolutionary Psychology and Human Behavior 19
#> 8                            scientometrics and bibliometrics research 17
#> 9                                    Personality Traits and Psychology 16
#> 10                            Scientific Computing and Data Management 16
#> 11                                    Social and Intergroup Psychology 16
#> 12                                Psychological Testing and Assessment 13
#> 13                            Advanced Statistical Modeling Techniques 12
#> 14                                  Death Anxiety and Social Exclusion 12
#> 15                           Personality Disorders and Psychopathology 10
#> 16                  Academic and Historical Perspectives in Psychology  9
#> 17                                     Cognitive Abilities and Testing  8
#> 18                              Statistical Methods in Clinical Trials  7
#> 19                      Psychological Well-being and Life Satisfaction  6
#> 20                                 Optimal Experimental Design Methods  5
#> 21                              Psychometric Methodologies and Testing  5
#> 22                             Advanced Statistical Methods and Models  4
#> 23                                 Complex Network Analysis Techniques  4
#> 24                                    Crime Patterns and Interventions  4
#> 25                               Functional Brain Connectivity Studies  4
#> 26                                     Health disparities and outcomes  4
#> 27                           Psychotherapy Techniques and Applications  4
#> 28                     Cognitive and psychological constructs research  3
#> 29                                Counseling Practices and Supervision  3
#> 30                                        Delphi Technique in Research  3
#> 31                               Evaluation and Performance Assessment  3
#> 32                                   Family Dynamics and Relationships  3
#> 33                             Forecasting Techniques and Applications  3
#> 34                            Neural and Behavioral Psychology Studies  3
#> 35                                   Psychology, Coaching, and Therapy  3
#> 36                              Religion, Spirituality, and Psychology  3
#> 37                            Sensory Analysis and Statistical Methods  3
#> 38                                        Technology and Data Analysis  3
#> 39                                          AI in Service Interactions  2
#> 40                             Advanced Statistical Process Monitoring  2
#> 41                                Behavioral and Psychological Studies  2
#> 42                                  Big Data and Business Intelligence  2
#> 43                                Biomedical and Engineering Education  2
#> 44                                          Child Welfare and Adoption  2
#> 45         Child and Adolescent Psychosocial and Emotional Development  2
#> 46                             Computational and Text Analysis Methods  2
#> 47                           Criminal Justice and Corrections Analysis  2
#> 48                                                Data Analysis with R  2
#> 49                                    Data Visualization and Analytics  2
#> 50                                    Digital Innovation in Industries  2
#> 51                                 Digital Mental Health Interventions  2
#> 52                                        Emotion and Mood Recognition  2
#> 53                              Emotional Intelligence and Performance  2
#> 54                                         Emotions and Moral Behavior  2
#> 55                                         Ethics in Clinical Research  2
#> 56                                      Evolution and Genetic Dynamics  2
#> 57                                     Face Recognition and Perception  2
#> 58                   Genetics, Bioinformatics, and Biomedical Research  2
#> 59               Health Systems, Economic Evaluations, Quality of Life  2
#> 60                                 Health and Medical Research Impacts  2
#> 61                                          Health and Medical Studies  2
#> 62                                  Libraries and Information Services  2
#> 63                                      Medical Research and Practices  2
#> 64                                      Misinformation and Its Impacts  2
#> 65                               Mobile Crowdsensing and Crowdsourcing  2
#> 66                                   Philosophy and History of Science  2
#> 67                             Psychiatry, Mental Health, Neuroscience  2
#> 68                    Psychological and Temporal Perspectives Research  2
#> 69                               Psychology Research and Bibliometrics  2
#> 70                          Psychology of Moral and Emotional Judgment  2
#> 71                  Psychopathy, Forensic Psychiatry, Sexual Offending  2
#> 72                           Qualitative Comparative Analysis Research  2
#> 73                            Sexual Assault and Victimization Studies  2
#> 74                                    Social Robot Interaction and HRI  2
#> 75                                     Sociology and Education Studies  2
#> 76                                    Sports Analytics and Performance  2
#> 77                          Statistical Methods and Bayesian Inference  2
#> 78                                 Academic Publishing and Open Access  1
#> 79                           Adolescent Sexual and Reproductive Health  1
#> 80                               Advanced Malware Detection Techniques  1
#> 81                               Advances in Oncology and Radiotherapy  1
#> 82  Anxiety, Depression, Psychometrics, Treatment, Cognitive Processes  1
#> 83                              Big Data Technologies and Applications  1
#> 84                                   Body Image and Dysmorphia Studies  1
#> 85                                          COVID-19 and Mental Health  1
#> 86                           Cardiac, Anesthesia and Surgical Outcomes  1
#> 87                                       Cognitive Science and Mapping  1
#> 88                                    Community Health and Development  1
#> 89                            Complex Systems and Time Series Analysis  1
#> 90                                   Conflict of Laws and Jurisdiction  1
#> 91                                        Corporate Governance and Law  1
#> 92                                 Corporate Governance and Management  1
#> 93                                         Data Analysis and Archiving  1
#> 94                             Data Mining Algorithms and Applications  1
#> 95                                         Data Quality and Management  1
#> 96                         Deception detection and forensic psychology  1
#> 97                            Decision-Making and Behavioral Economics  1
#> 98                                        Digital Media and Visual Art  1
#> 99              Diverse Approaches in Healthcare and Education Studies  1
#> 100                                E-Learning and Knowledge Management  1
#> 101                             Educational Assessment and Improvement  1
#> 102                                 Educational Games and Gamification  1
#> 103                          Educational Strategies and Epistemologies  1
#> 104                               Environmental Science and Technology  1
#> 105                                   Ethics in Business and Education  1
#> 106                                         Ethics in medical practice  1
#> 107                           Evolutionary Game Theory and Cooperation  1
#> 108                          Experimental Behavioral Economics Studies  1
#> 109                          Explainable Artificial Intelligence (XAI)  1
#> 110                                          Family Support in Illness  1
#> 111                             Family and Disability Support Research  1
#> 112                                Fault Detection and Control Systems  1
#> 113                                   Gene Regulatory Network Analysis  1
#> 114                              Grief, Bereavement, and Mental Health  1
#> 115                            Hate Speech and Cyberbullying Detection  1
#> 116                                 Health, psychology, and well-being  1
#> 117                                         Human Motion and Animation  1
#> 118               Innovative Teaching Methodologies in Social Sciences  1
#> 119                   Intergenerational Family Dynamics and Caregiving  1
#> 120                                    Korean Urban and Social Studies  1
#> 121                           Leadership, Courage, and Heroism Studies  1
#> 122                                   Linguistic research and analysis  1
#> 123                                         Media Influence and Health  1
#> 124                       Model-Driven Software Engineering Techniques  1
#> 125                              Motivation and Self-Concept in Sports  1
#> 126                                 Neural dynamics and brain function  1
#> 127                              Opinion Dynamics and Social Influence  1
#> 128                                     Optimism, Hope, and Well-being  1
#> 129                                 Paranormal Experiences and Beliefs  1
#> 130                                           Plant and animal studies  1
#> 131                        Psychiatric care and mental health services  1
#> 132                           Reliability and Agreement in Measurement  1
#> 133                                  Religion and Society Interactions  1
#> 134                                 Religion, Society, and Development  1
#> 135                               Reproductive Health and Technologies  1
#> 136                                   SARS-CoV-2 detection and testing  1
#> 137                                        Sex work and related issues  1
#> 138                                Sexuality, Behavior, and Technology  1
#> 139                                        Sleep and related disorders  1
#> 140                                        Social Capital and Networks  1
#> 141                                          Social Media and Politics  1
#> 142                                   Social Power and Status Dynamics  1
#> 143                                      Software Engineering Research  1
#> 144                          Software Testing and Debugging Techniques  1
#> 145                                   Spatial Cognition and Navigation  1
#> 146                                       Sports Science and Education  1
#> 147                                      Stress Responses and Cortisol  1
#> 148                                  Teaching and Learning Programming  1
#> 149                                      Treatment of Major Depression  1
#> 150                             Winter Sports Injuries and Performance  1
#> 151                              Youth Education and Societal Dynamics  1
#> 
#> $interdisc_string
#> [1] "4 primary fields: Psychology (128), Decision Sciences (43), Computer Science (28), Health Professions (7). The evenness (ranging from 0=only one field to 1=even distribution among all fields) is 0.56."
#> 
# Analyze the network of a specific author for selected works
get_network(author.id = 'https://openalex.org/A5022479713', min_edges=1,
    doi = c(
        'https://doi.org/10.1037/pspp0000428',
        'https://doi.org/10.1017/S0033291722003294',
        'https://doi.org/10.5964/ps.6029',
        'https://doi.org/10.1146/annurev-psych-020821-114157'))
#> Error in get_network(author.id = "https://openalex.org/A5022479713", min_edges = 1,     doi = c("https://doi.org/10.1037/pspp0000428", "https://doi.org/10.1017/S0033291722003294",         "https://doi.org/10.5964/ps.6029", "https://doi.org/10.1146/annurev-psych-020821-114157")): unused argument (min_edges = 1)