Get a reference set for field normalized citation rates

This function downloads a large set of publications from a certain field and certain years. Retrieving 10,000 documents from the API takes about 1-2 min. Ideally, this should be cached locally.

Usage

get_reference_set(
  years,
  n_per_year = 10000,
  concept.id = "C15744967",
  verbose = TRUE
)

Arguments

years

The year(s) from which a sample of the reference field should be retrieved.

n_per_year

The number of documents to retrieve per requested year (optional, defaults to 10000). Values larger than 10000 are possible (they are split up to multiple OA requests).

concept.id

A vector of concept.ids to search for (optional, defaults to "C15744967", i.e. "Psychology")

verbose

Show OA API progress?

Value

A data frame containing the document id, year, cited_by_count, and number of authors of the retrieved documents

Examples

# Get reference set for "Psychology" for multiple years (small n here for demo)
psych_ref <- get_reference_set(
  2013:2020, n_per_year = 100,
  concept.id = "C15744967"
 )
#> [1] "2023-05-24 15:24:56.658033: Retrieving documents for 2013 ..."
#> Requesting url: https://api.openalex.org/works?filter=concept.id%3AC15744967%2Ctype%3Ajournal-article%2Cfrom_publication_date%3A2013-01-01%2Cto_publication_date%3A2013-12-31%2Cis_paratext%3Afalse%2Cis_retracted%3Afalse%2Cauthors_count%3A%3E0%2Chas_doi%3Atrue&sample=100
#> Getting 1 page of results with a total of 100 records...
#> [1] "Retrieved 100 documents"
#> [1] "2023-05-24 15:24:57.943126: Retrieving documents for 2014 ..."
#> Requesting url: https://api.openalex.org/works?filter=concept.id%3AC15744967%2Ctype%3Ajournal-article%2Cfrom_publication_date%3A2014-01-01%2Cto_publication_date%3A2014-12-31%2Cis_paratext%3Afalse%2Cis_retracted%3Afalse%2Cauthors_count%3A%3E0%2Chas_doi%3Atrue&sample=100
#> Getting 1 page of results with a total of 100 records...
#> [1] "Retrieved 100 documents"
#> [1] "2023-05-24 15:24:59.103499: Retrieving documents for 2015 ..."
#> Requesting url: https://api.openalex.org/works?filter=concept.id%3AC15744967%2Ctype%3Ajournal-article%2Cfrom_publication_date%3A2015-01-01%2Cto_publication_date%3A2015-12-31%2Cis_paratext%3Afalse%2Cis_retracted%3Afalse%2Cauthors_count%3A%3E0%2Chas_doi%3Atrue&sample=100
#> Getting 1 page of results with a total of 100 records...
#> [1] "Retrieved 100 documents"
#> [1] "2023-05-24 15:25:00.392592: Retrieving documents for 2016 ..."
#> Requesting url: https://api.openalex.org/works?filter=concept.id%3AC15744967%2Ctype%3Ajournal-article%2Cfrom_publication_date%3A2016-01-01%2Cto_publication_date%3A2016-12-31%2Cis_paratext%3Afalse%2Cis_retracted%3Afalse%2Cauthors_count%3A%3E0%2Chas_doi%3Atrue&sample=100
#> Getting 1 page of results with a total of 100 records...
#> [1] "Retrieved 100 documents"
#> [1] "2023-05-24 15:25:01.58352: Retrieving documents for 2017 ..."
#> Requesting url: https://api.openalex.org/works?filter=concept.id%3AC15744967%2Ctype%3Ajournal-article%2Cfrom_publication_date%3A2017-01-01%2Cto_publication_date%3A2017-12-31%2Cis_paratext%3Afalse%2Cis_retracted%3Afalse%2Cauthors_count%3A%3E0%2Chas_doi%3Atrue&sample=100
#> Getting 1 page of results with a total of 100 records...
#> [1] "Retrieved 100 documents"
#> [1] "2023-05-24 15:25:02.843148: Retrieving documents for 2018 ..."
#> Requesting url: https://api.openalex.org/works?filter=concept.id%3AC15744967%2Ctype%3Ajournal-article%2Cfrom_publication_date%3A2018-01-01%2Cto_publication_date%3A2018-12-31%2Cis_paratext%3Afalse%2Cis_retracted%3Afalse%2Cauthors_count%3A%3E0%2Chas_doi%3Atrue&sample=100
#> Getting 1 page of results with a total of 100 records...
#> [1] "Retrieved 100 documents"
#> [1] "2023-05-24 15:25:04.129693: Retrieving documents for 2019 ..."
#> Requesting url: https://api.openalex.org/works?filter=concept.id%3AC15744967%2Ctype%3Ajournal-article%2Cfrom_publication_date%3A2019-01-01%2Cto_publication_date%3A2019-12-31%2Cis_paratext%3Afalse%2Cis_retracted%3Afalse%2Cauthors_count%3A%3E0%2Chas_doi%3Atrue&sample=100
#> Getting 1 page of results with a total of 100 records...
#> [1] "Retrieved 100 documents"
#> [1] "2023-05-24 15:25:05.393481: Retrieving documents for 2020 ..."
#> Requesting url: https://api.openalex.org/works?filter=concept.id%3AC15744967%2Ctype%3Ajournal-article%2Cfrom_publication_date%3A2020-01-01%2Cto_publication_date%3A2020-12-31%2Cis_paratext%3Afalse%2Cis_retracted%3Afalse%2Cauthors_count%3A%3E0%2Chas_doi%3Atrue&sample=100
#> Getting 1 page of results with a total of 100 records...
#> [1] "Retrieved 100 documents"