Get a reference set for field normalized citation rates

This function downloads a large set of publications from a certain field and certain years. Retrieving 10,000 documents from the API takes about 1-2 min. Ideally, this should be cached locally.

Usage

get_reference_set(
  years,
  n_per_year = 10000,
  concept.id = "C15744967",
  type = "article",
  verbose = TRUE,
  seed = NULL,
  save_intermediate = NULL
)

Arguments

years

The year(s) from which a sample of the reference field should be retrieved.

n_per_year

The number of documents to retrieve per requested year (optional, defaults to 10000). Values larger than 10000 are possible (they are split up to multiple OA requests).

concept.id

A vector of concept.ids to search for (optional, defaults to "C15744967", i.e. "Psychology")

type

Reference sets should refer to the same type of publication; defaults to "article"

verbose

Show OA API progress?

seed

Set a seed for reproducible analyses. However, as the underlying OA database changes frequently, the results will still not be very stable ...

save_intermediate

If a path is provided here, the intermediate downloaded files are saved at that path.

Value

A data frame containing the document id, year, cited_by_count, and number of authors of the retrieved documents

Examples

# Get reference set for "Psychology" for multiple years (small n here for demo)
psych_ref <- get_reference_set(
  years = 2018:2020, n_per_year = 20,
  concept.id = "C15744967"
 )
#> [1] "2025-03-07 09:47:09.891046: Retrieving documents for 2018 ..."
#> Requesting url: https://api.openalex.org/works?filter=concept.id%3AC15744967%2Ctype%3Aarticle%2Cfrom_publication_date%3A2018-01-01%2Cto_publication_date%3A2018-12-31%2Cis_paratext%3Afalse%2Cis_retracted%3Afalse%2Cauthors_count%3A%3E0%2Chas_doi%3Atrue%2Cprimary_location.source.has_issn%3Atrue&sample=20&select=id%2Cauthorships%2Cpublication_year%2Ccited_by_count
#> Getting 1 page of results with a total of 20 records...
#> [1] "Retrieved 20 documents"
#> [1] "2025-03-07 09:47:11.042006: Retrieving documents for 2019 ..."
#> Requesting url: https://api.openalex.org/works?filter=concept.id%3AC15744967%2Ctype%3Aarticle%2Cfrom_publication_date%3A2019-01-01%2Cto_publication_date%3A2019-12-31%2Cis_paratext%3Afalse%2Cis_retracted%3Afalse%2Cauthors_count%3A%3E0%2Chas_doi%3Atrue%2Cprimary_location.source.has_issn%3Atrue&sample=20&select=id%2Cauthorships%2Cpublication_year%2Ccited_by_count
#> Getting 1 page of results with a total of 20 records...
#> [1] "Retrieved 20 documents"
#> [1] "2025-03-07 09:47:12.26803: Retrieving documents for 2020 ..."
#> Requesting url: https://api.openalex.org/works?filter=concept.id%3AC15744967%2Ctype%3Aarticle%2Cfrom_publication_date%3A2020-01-01%2Cto_publication_date%3A2020-12-31%2Cis_paratext%3Afalse%2Cis_retracted%3Afalse%2Cauthors_count%3A%3E0%2Chas_doi%3Atrue%2Cprimary_location.source.has_issn%3Atrue&sample=20&select=id%2Cauthorships%2Cpublication_year%2Ccited_by_count
#> Getting 1 page of results with a total of 20 records...
#> [1] "Retrieved 20 documents"