Compute h-index and hIa-index

TODO: Could use a time-window (currently, it uses all citations since publication).

Usage

h_index(
  search = NULL,
  display_name = NULL,
  author.id = NULL,
  first_pub_year = NA,
  academic_age_bonus = 0
)

Arguments

search

A character string with the author's name for searching in the OpenAlex API. Allows fuzzy search (e.g. with or without middle initial).

display_name

A character vector with the exact author's names for searching in the OpenAlex API. You cvan provide multiple variants in a vector.

author.id

A character vector with OpenAlex author ID(s) (i.e., an ID starting with "A.......") for searching in the OpenAlex API.

first_pub_year

An integer with the year of the author's first publication. If not provided, it is estimated from the retrieved publications.

academic_age_bonus

Years that are subtracted from active academic life (e.g. due to child care). For example, SNF subtracts 1.5 years per child (see https://www.snf.ch/en/cciM9NWuvhOVKRxv/news/news-200803-career-funding-the-researchers-overall-performance-counts).

Value

A list containing the net academic age, h-index and hIa-index of the author.

Details

Computes the h-index and hIa-index of an author. You need to provide exactly one of the parameters search, display_name or author.id:

search allows fuzzy search of the author's name (e.g. with or without middle initial). Might return too many matches

display_name requires an exact match but allows to provide a vector with multiple versions of a name (e.g., c("Hans Müller", "Hans Mueller", "Hans J. Muller"))

author.id is the OpenAlex ID for authors.

Often the same author is listed under different IDs in the OpenAlex database, therefore these are merged by the function. But what if the function includes wrong authors? Then you need to manually screen the table with author names, record the IDs from the correct aliases of the target person (starting with "A.......") and enter these author IDs into the author.id parameter.

Examples

h_index(search = "Felix Schönbrodt")
#> # A tibble: 5 × 3
#>   display_name        works_count cited_by_count
#>   <chr>                     <int>          <int>
#> 1 Felix D. Schönbrodt         245           7200
#> 2 NA                           NA             NA
#> 3 NA                           NA             NA
#> 4 NA                           NA             NA
#> 5 NA                           NA             NA
#> Computation of academic age (relative to first publication):
#> Retrieving year of first publication from data: 2006. Is that plausible?
#> 
#> $net_academic_age
#> [1] 18
#> 
#> $h_index
#> [1] 27
#> 
#> $h_Ia
#> [1] 0.8888889
#> 
h_index("Markus Bühner", first_pub_year = 1997) # adjust year of first publication
#> # A tibble: 5 × 3
#>   display_name  works_count cited_by_count
#>   <chr>               <int>          <int>
#> 1 Markus Bühner         205           5306
#> 2 NA                     NA             NA
#> 3 NA                     NA             NA
#> 4 NA                     NA             NA
#> 5 NA                     NA             NA
#> $net_academic_age
#> [1] 27
#> 
#> $h_index
#> [1] 36
#> 
#> $h_Ia
#> [1] 0.4814815
#>